Factored middleware into its own file for users
